Refactored to use time correctly

+++ src/cmd/models/measurements.go
@@ -3,8 +3,8 @@ package models
 import "time"
 
 type Measurements struct {
-  Id        int         `json:"id"`
-  UserId    int         `json:"user_id"`
+  Id        int64       `json:"id"`
+  UserId    int64       `json:"user_id"`
   Weight    float64     `json:"weight"`
   Height    float64     `json:"height"`
   BodyFat   float64     `json:"body_fat"`
